Move `flow_ops.rs` code to `scope.rs`
Summary:
Flow operations are scope operations. It's much simpler if they have access
to scope internals, and if making that so does not require exposing scope
internals to the rest of Pyrefly.

Now nontrivial scope refactors - like changing how flow merging works in
fundamental ways that also require updating datatypes - can be done as a
scope-only change, respecting public interfaces.

I was torn about this for a while because it makes `scope.rs` bigger, but I
concluded that actual modularity is far more important than module length
as a principle for code organization - scope is actually complex, so it can
be a very big module; that's much better than a complex and slightly smaller
module where the implementation details have to be public just to shorten
the file length.

Reviewed By: grievejia

Differential Revision: D83055326

fbshipit-source-id: 40aa3655c28770f61f405f7d199bd14f8e36b10f