Update comment on `bind_name`
Summary:
The bit about the default is out of date, I removed the unnecessary
`Default` bindings and simplified this funciton but missed part of the
doc comment.

I also think it's helpful to provide a hint about why we're returning
the `Idx<KeyAnnotation>`, which isn't obvious if you don't already
know the answer.

Reviewed By: grievejia

Differential Revision: D83069635

fbshipit-source-id: bf0a2673ae73d2b16e816a1ff20ec302b57fc3c1