Special-case function calls to `prod_assert()` with 1 or 2 arguments
Summary:
Instagram uses a custom function `util.prod_assert()` to replace the `assert` statement. It would be really nice if Pyrefly could treat this function in similar way as a normal assertions (to get all the narrowing behaviors that we coded into assertions).

The easiest way we could approach this issue is to hard-code support for it within the type checker. This diff took that approach, and treat any function calls whose target is the name `prod_assert` and the number of arguments is 1 or 2 as an assertion. It is true that this might be disruptive to projects which defines the name `prod_assert` in alternative ways, but (1) it would be rather rare for projects to name a function `prod_assert` without making it an actual assertion (if we are concerned about the naming I'd be happy to codemod it to some other names that's even less likely get a conflict, e.g. `ig_prod_assert`), and (2) we've already done this kind of name-based special casing elsewhere. E.g. `self.assertIsNotNone()` and friends within unittests are all special-cased via function names.

The alternative approach is to codemod IG and replace all invocations to
```
prod_assert(isinstance(XXX, YYY))
```
into
```
if not isinstance(XXX, YYY):
  raise prod_assert(False)
```
which may functionally work but may be a hard sell to Instagram devs.

Reviewed By: stroxler

Differential Revision: D83007946

fbshipit-source-id: 47c8be21e77f7e0c28242dab0193e69a41becb65