Implement sourcedb lookup
Summary:
This diff adds one of the remaining functions for a sourcedb for `BuckSourceDatabase`, `lookup`.

Lookup does a breadth-first search of the module we're trying to import, starting at the importing module's target, and continuing through its dependencies, returning the `ModulePath` for the first `ModuleName` that matches.

If this becomes a performance bottleneck, we can do more to condense the sourcedb or move to a single config per target approach.

One design choice made here is that we always select the target used for a path lookup alphabetically, rather than selecting the first-opened target. Jia raises concerns around this, which I'll make a post on soon, and we can revisit/adjust the implementation once we come to a conclusion.

Reviewed By: grievejia

Differential Revision: D82499623

fbshipit-source-id: d78d1d9e617579893ae704ae9e32d9fc3ec15011