Implement sourcedb handle_from_module_path
Summary:
This diff implements `handle_from_module_path` for `BuckSourceDatabase`, which finds the associated target, if it exists, and creates a handle for it.

This function is used for creating handles at the start of a type check only, and handles that are transitively created inherit their sys info from their importing handle.

Reviewed By: grievejia

Differential Revision: D82601222

fbshipit-source-id: 498e133a40612d1d0e58fe101e090b91c87e6a37