Move locking into sourcedb
Summary:
This diff moves locking functionality into the sourcedb, so we can be more granular about when we decide to lock things (avoid write-locking when a read will do, like when we're figuring out if we need to do a sourcedb update).

Since the plan is to invalidate the find cache after an update, this should be fine to do.

Reviewed By: grievejia

Differential Revision: D82499627

fbshipit-source-id: 594c74e3a0358ca09fa4622e5796ff1d3adc4905