Enable build systems in IDE and disable whole project type checks
Summary:
This diff removes the printed error from trying to run build system type checking in the IDE.

We also output an error when trying to check a build system without files, since that should be heavily discouraged in most cases. In the future, we can come back to making it configurable.

Reviewed By: SamChou19815

Differential Revision: D82733064

fbshipit-source-id: 5d2ad47fec4044a9c9b3d8f76e2eeade86f8243c