Tweak the get-lookup-key function
Summary:
It's confusing to call this `key()` because that could easily mean
"the `Key` associated with the binding for this thing".

It's also kind of confusing to invent a fake `Name`, this isn't
a valid AST name at all. We could use a custom datatype, but i think
String seems okay given that there are only a few relevant places
anyway.

Reviewed By: rchen152

Differential Revision: D83156244

fbshipit-source-id: ef12bb9700dedbe7839ad13de0d1ddd24bb0e0ea