Avoid callbacks for making legacy tparam bindings
Summary:
By using a flat match, we can avoid callbacks. I think this is slightly easier
to read as is - the code is more concise, the doc comment can describe more
complexity, and we don't have indirection (or the reduced IDE navigation abilitites
that result from it) coming from callbacks.

But my main motivation is that now `lookup_legacy_tparam_from_idx` is simplified
enough (because we pushed some of the fiddly bits down) that it's easy to add
some complexity there in order to make the bindings in a smarter way.

Reviewed By: rchen152

Differential Revision: D83156245

fbshipit-source-id: cec00601f4caf03c709a047d57412ca4d0a5f9b2