Re-structure `lookup_legacy_tparam_from_idx`
Summary:
Instead of a for loop with a `continue`, use a while loop with a counter and
early return to get the same behavior.

This change is probably roughly neutral on its own - I think the structure of
the code expresses intent clearer when we write the loop as a graph traversal
with early exit using a glorified goto instead of vice/versa, but this
is debateable; we now have to write down the get call twice and do some more
fiddling with the `counter`.

But if we add more processing to the result, I think this will be much nicer
because instead of the pipeline starting with `make_legacy_tparam` being
buried in a loop, it comes below the loop - so the graph traversal embodied
in the loop is no longer visually coupled to whatever we do with the result.

It's easier (at least for my brain) to read that and see what's going on.

Reviewed By: rchen152

Differential Revision: D83156256

fbshipit-source-id: 7e348ef5dc201f0ebe8bd6875e56ad05e1d40955