Swap function order
Summary:
It's easier to read the stack of layers here (intercept_lookup ->
lookup_legacy_tparam -> lookup_legacy_tparam_from_idx -> make_legacy_tparam)
if they come one right after the other.

And it's a little easier to ready the story told by the code if `add_name_definitions`
is at the bottom, because it executes after everything in the stack above.

Reviewed By: rchen152

Differential Revision: D83156248

fbshipit-source-id: 623b3607153f1a3c36dbab21fb3bca430e27ab6c