Move `test_bool_special_exports_bug`
Summary:
It's not at all obvious what this test does just from looking at
it, but it's actually a very important test because
(a) we handle legacy tparams and possible legacy tparams incorrectly
    in scopes - for example, mutable capture lookups will incorrectly
    treat our made-up bindings as actual names in runtime scope, which
    is not correct.
(b) In the process of this we also have to do questionable
    things with flow style, because we have to pretend that these fake
    names in the parameter scope (which is a real scope but only contains
    PEP-695 type parameters - none of these legacy things exist in a real
    scope at runtime) have flow values.
(c) attempts to either address (a) or just deal with the fact that (b) gets
    in the way of refactors will frequently break this test

I've now run into this problem and spent quite a while staring at the code,
which is hard to understand both because it is actually complicated and because
the data model is actually incorrect relative to the runtime.

It seems worth explaining what the test does, in enough detail to aid with
debugging. I try to keep the explanation vague enough that refactors shouldn't
lead to it going out of date.

It also ought to live in the legacy generic test suite, because the thing
being tested isn't `bool` per se, but rather whether our scope mangling
for legacy generics breaks special export lookups.

Reviewed By: rchen152

Differential Revision: D83156249

fbshipit-source-id: 3809afdcb09ab69e5de7c80bacd054f165096337