Rename the collector datatype
Summary:
Now that there's no inversion of control, the name `Builder` feels awkward,
it implies some kind of complex architecture that we don't have now.

Rename it to "collector" which I think better captures what's really going
on, it's just a simple data structure with a few methods to help collect
and retrieve the legacy params that the `BindingsBuilder` finds.

Reviewed By: rchen152

Differential Revision: D83156246

fbshipit-source-id: f94512039b623df0e4b14dbd4a7335cad735904d