Move code around
Summary:
After talking with Rebecca, we concluded that the `LegacyTParamBuilder` pattern
is unnecessarily confusing: the `BindingsBuilder` holds one of these in scope
and passes itself into methods that can call back, which produces a flip-flop
of who owns what.

None of this actually matters for what the code *does* (and as a result it's
trivial to eliminate), but it does matter for how the code is organized and how
easy it is to navigate, it leads to an inversion-of-control feeling that
reminds me of confusing Spring Boot applications.

We decided to get rid of this pattern, by taking the exact same code and making
the current `TParamsBuilder` just be plain data - anything that needs access to
the `BindingsBuilder` should just be a normal `BindingsBuilder` method, like
we do everywhere else in bindings.

This diff is the first half of making that change:
- I split the `BindingsBuiler` into two blocks, one for the "core"
  functionality and one for the legacy-tparam-specific code.
- I move the `LegacyTParamId` and `LegacyTParamBuilder` code above the second
  block
- I moved the `lookup_keys` method of `LegacyTParamBuilder` up, to sit underneath
  `new`.

In the next diff, I'll move all of the `LegacyTParamBuilder` methods *except*
for `new` and `lookup_keys` down, which will get rid of the inversion of control.

Reviewed By: rchen152

Differential Revision: D83156242

fbshipit-source-id: d97025bc0a6a56692cb94937beacb31bfb5fde01