Add CLI help message to discourage the use of `--watch`
Summary:
great_experiment

I still think watch mode would be useful but the sad reality is that it has not been that well-tested and did not see much usage in practice. I've marked it as "highly experimental" in this diff and wouldn't mind if anyone wants to remove it later down the road.

Reviewed By: ndmitchell

Differential Revision: D83117729

fbshipit-source-id: 607324ee7f66ccf13be07549d04f8eccffcbaaa9