Track a bug in module-atter legacy tparam handling
Summary:
The test case describes the bug pretty clearly. This might actually cause some major
confusion for users, it's probably somewhat important to address in the moderately
near future.

This test is also useful as a regression test of existing code - an earlier version of
D83156252 had a panic that wasn't caught by our test suite, but is with this test in the mix

Reviewed By: rchen152

Differential Revision: D83177891

fbshipit-source-id: 976efffe99fcb0dc68d95a7f577aa3e046e5d6f8