Record bad behavior due to legacy-tparam-in-static
Summary:
This is the main case I know of where the current layout of `Scope`, where
possible legacy type parameters are put into `Static` and therefore shadow
the true definitions, is broken: mutable captures will find the wrong definition,
leading to false positives.

I don't plan to fix it soon or think this is very important (how often are you
going to mutably capture a name that was used in the annotations of the
function header? Mutable captures are already rare, and this is very strange)
but I do think it's a flaw in our architecture worthy of being tracked by
a unit test.

The solution would probably be to move possible-legacy-tparam bindings into a
special place that is neither `Static` nor `Flow` - one option is to put them
in the `ScopeKind` enum variant for parameter scopes, similar to how we track
self asignments in the class `ScopeKind` variant - and implement special lookup
rules so that only static type lookups are affected and value lookups correctly
see the actual runtime binding.

Reviewed By: rchen152

Differential Revision: D83172320

fbshipit-source-id: 2e2cdc02928744b2637fc7eea56f90b55cc8945c