Legacy tparams are not flow / Definition
Summary:
**About Definition**

It's a little ambiguous whether an actual legacy type parameter should
be treated as a go-to-def target by the IDE, e.g. if I have
```
T = TypeVar("T")
def f(x: T):
    y: T = x
```
then there are two questions:
- If I hover on the `T` after `x`, where should I go? Almost certainly
  the answer is to the `T = TypeVar("T")`
- If I hover on the `T` after `y`, where should I go? I think either
  answer is defensible; if the function is huge then going to the
  `T` after `x` is arguably nice.

But Pyrefly today isn't smart enough to do this regardless of what we
want, because lots of things that *aren't* legacy type parameters,
including any forward reference or imported name, also gets a
binding as a potential legacy type param.

That means that we could get a lot of go-to-def results wrong! It seems
better for now to never treat the tparam as a definition. This undoes
the behavior change from the previous diff.

This also is sort of nice as a BE win, since legacy type parameters are
confusing enough that having a dedicated `Key` seems good for both
understanding the rust code and easier debugging of bindings.

**About Flow**

A possible legacy tparam is unambiguously not a flow value at all,
the runtime value is exactly whatever it was in the first place.

The fact that we were putting this in flow actually made it hard to
avoid some kinds of bugs - for example if we didn't copy over the
flow style from the parent, we would break special export resolution.

This diff makes it so that the re-baound name is only defined in the static
scope, not the flow scope, eliminating the need to copy flow info. This
is important mainly because I want to make changes to how we model narrows,
and the copying of flow info was in my way... this change is in fact
the whole reason I started looking at legacy tparams, because copying the
flow info made no sense and I couldn't figure out a way not to block my
refactor on some kind of ugly hack here.

**About Static**

It's not really quite right to put the name in a scope at all, even
in just `Static`, because scope is modeling mainly *runtime* behavior,
and this name is 100% not part of any actual runtime scope - a few
edge cases don't work when we pretend it is.

I'll stack a diff with a bug illustrating why this can cause problems,
with a summary suggesting a potential solution. But I don't think the
bugs are very urgent, this seems good enough for now and my flow-merging
work is no longer blocked.

In addition to that problem, there's another preexisting issue that
becomes more obvious with this change: we're potentially producing a bunch
of separate bindings that might narrow attributes on an imported module.
This is a problem because then only one of the narrows can be in scope
at a time. I'll stack a test case for that as well; the symptom we see
here is that `Static::upsert` has to handle it by picking who wins.

Reviewed By: rchen152

Differential Revision: D83156252

fbshipit-source-id: 048b88a9a22ec1d3bcaf1aa1f89f110cf94a6a95