Separate parameter var solving from other vars
Summary:
Context:

I am working toward new generic instantiation, where we can consider multiple upper/lower bounds when determining the instantiation of a typevar.

This diff:

Distinguish "Variable::Parameter" which is unlike other vars, and is solved in a simpler way.

Function parameters are special -- we solve them in exactly one place. We don't expect them to ever appear in a constraint.

Lambda parameters are also special, but not in quite the same way. We expect them to appear in constraints -- specifically the constraint emitted by decompose_lambda, similar to other decompose operations, which use Variable::Unwrap.

So, this diff moves the solving of function parameters to a dedicated function & asserts that these vars never appear in a constraint.

I also change lambda params to be Unwrap, since that is how they are used. Unwrap will accumulate bounds, and will be solved after the decompose operations.

Reviewed By: stroxler

Differential Revision: D83170765

fbshipit-source-id: bb028d4c48930f0afc88b36afe08411a60a2bd67