Re-structure `lookup_legacy_tparam` a bit
Summary:
I'm about done making improvements to legacy-tparam logic for now; the last
thing I really want to fix before I move on is that the use of an
`Either` type wrapping an `Option` leads to verbose annotations that aren't
very self-explanatory, with double-nesting that's a little hard to follow.

I want to make an enum for this, but the current code structure isn't very
friendly to getting rid of the nesting - the use of `found` downstream
of a `map` on itself is both pretty weird looking and tightly coupled to
using an `Either<Option<...>>` rather than a flat type to express the cases.

So I'm tweaking the control flow in this diff (no change to behavior or types)
and I'll stack a diff introducing an enum for the lookup result.

Reviewed By: rchen152

Differential Revision: D83180893

fbshipit-source-id: a2c115bb4cf5a040ce62c51c5e7ea5815a64b47b