Change class body FlowStyle::Uninitialized semantics
Summary:
A `FlowInfo::Uninitialized` can happen in several cases - for example if a name
was declared in a bare `AnnAssign` with no value or was defined but then
deleted.

If we run into this case in a class body - which has dynamic scope that falls
back to enclosing scopes when the name isn't defined - we should keep looking.

Fixes https://github.com/facebook/pyrefly/issues/1146

Reviewed By: rchen152

Differential Revision: D83281065

fbshipit-source-id: 60608fbfac6e4cb1e6bdda27e68a145bcff14466