Don't mark possible tparams as `IsInitialized::Maybe`
Summary:
This is a weird thing to do - it kind of makes sense in the sense
that we don't know for sure, but in this case we wouldn't want to
raise false positives.

In addition, it turns out it doesn't matter anyway - `intercept_lookup`
will only be used in static-type-style lookups, which don't check
initialization anyway.

Reviewed By: rchen152

Differential Revision: D83212553

fbshipit-source-id: bc6cb781d7c3bacf639ddfe61175e289b34269d1