Rename `Initialized`
Summary:
This type is in a big namespace where it's meaning isn't clear; in fact, there's
an `IsInitialized` in `bindings.rs` that means something else - whether a name
is initialized in the flow (I'll rename that in the next diff).

This name makes it much clearer what we mean, and the comment lets us hint at
why this matters... I had to dig around a bit because it's not obvious.

Reviewed By: rchen152

Differential Revision: D83204963

fbshipit-source-id: 4fd05c713f3df24b5235a4a390ed075b5fca1f0f