Add a test of various class-body-field edge cases
Summary:
I need to make some changes to how Pyrefly handles AnnAssign with
no value in flow, and therefore also how Pyrefly treats class attributes
that aren't defined in the flow at the end of the class body.

A few of the edge cases are ambiguous and I don't actually care much what
we do (it's impossible to statically determine the right signature
if there's highly dynamic logic where an attribute may or may not exist, so
what we do in non-analyzable cases is pretty arbitrary) but I do think we
need test coverage showing any behavior changes, and also ensuring that we
don't break any of the cases that really need to work.

Reviewed By: rchen152

Differential Revision: D83281066

fbshipit-source-id: 75fd4afd6b3deb55470fa19413c82e9a6a6613be