Make `Usage::Narrowing` carry an optional idx
Summary:
See attached github issue for context.

Basically, if we're in another binding (which ought to be the first use), then in a sub-expression we enter the narrow and we start treating it as something downstream of the first use, we would creates a cycle. For example,

```
y = []
x = y if len(y) == 0 else [5]
```
We want the `NameAssign(x)` binding to be the first use of the raw `y` and then have `Pin(y)` depend on it, but we also create a `Narrow(y @ 'len(y) == 0')` binding, which is depending on `Pin(y)`, and now we have a cycle.

D76693316 is actually an attempt to fix the same problem, but the solution only works if no narrowing is involved. This diff basically extends the pre-existing mechanism, insert the same bookkeeping info to `Usage::Narrow`, and mimic the behavior of using the raw type instead of the pinned type upstream.

This is *not* a 100% foolproof solution. Using raw types more often would "open up" more chances for unpinned type to expose to downstream code, leading to potential non-determinism. E.g.
```

def f(x: list[int]) -> int: ...
def g(x: list[str]) -> str: ...

def test():
    y = []
    x = (z := f(y), w := g(y)) if len(y) > 0 else None
    return w
```
Here depending on which expression gets calculated first `f(y)` vs. `g(y)` we would get different type errors. Given how walrus operators is rarely used in practice I'm less worried about it and thinks we should fix the issue at hand first. But I also want to emphasize that the risk / issue is not fully resolved, and we may want to be prepared to either backout this diff (if nondet affects our rollout) or future followups on better solutions.

Fixes https://github.com/facebook/pyrefly/issues/1030

Reviewed By: stroxler

Differential Revision: D83113229

fbshipit-source-id: 9e08a4a7952ce721d769af74a100a3cdac7577e1