Track a flow-merging bug in a simpler case
Summary:
This bug is "tracked" by the test case underneath, but deailing
with boolean operations is complicated and there could be other
bugs in the way.

It's good to have a simple if/else testing this.

I wrote a simple fix for this, but it turns out not to work
well with at least some kinds of merging (for example unreachable
patterns start throwing errors on every line), so we probably
need to live with it for a while - this is nowhere near the most
important flow merging issue right now.

Reviewed By: grievejia

Differential Revision: D83349999

fbshipit-source-id: e24c76cddf37a6ede4e734bc4eb1c042b5d2f579