Add a test of annotated legacy type vars
Summary:
These have special handling in `stmt.rs`, and I wanted to
know what the impact of an annotation is.

It breaks the handling; this seems like a bug and worthy
of having tracked in the tests, although it's not high priority
given that I've never seen someone actually do this.

Reviewed By: grievejia

Differential Revision: D83349995

fbshipit-source-id: e29256fdae16e55fa800c3f26644677761d66830