Split the upsert API into assign vs narrow
Summary:
This is maybe a tiny BE win as is; the new functions have simpler
APIs and express intent more clearly. But this is also strictly needed
if we want scope to eventually be able to distinguish assigned vs narrowed
types, which we are planning to use for improving Phi behaviors.

Reviewed By: grievejia

Differential Revision: D83349993

fbshipit-source-id: 334f78b1ed1ae628b4c48dee3ddac471caa284b5