Inline `class_field_in_body` definition
Summary:
At this point `ClassFieldInBody` is not really helpful - it was a
necessary abstraction back when `class.rs` was still responsible
for constructing `ClassFieldDefinition`s, but now that we've moved
that to `scope.rs` we can just directly get what we need.

That will be better both because it's simpler and because it's much
easier to update a one-hop transformation when we change `FlowInfo`
than to update a two-hop transformation.

I'm starting by inlining the existing logic, so that when I eliminate it
in the next diff it will be easy to see the change in-place.

Reviewed By: grievejia

Differential Revision: D83349994

fbshipit-source-id: ee053aafd9e446edd02413d6d228c8854a2630a6