Remove `bug` tag and describe now-passing test in more detail
Summary:
I missed removing the `bug` tag in a refactor a few months ago.

Also, I think a comment is warrented to explain what it is we're testing (an
early implementation of looping stored loops on `Scopes` rather than `Scope`,
which led to us incorrectly marking entire function top-levels as "in a loop"
for flow-merging purposes).

Reviewed By: yangdanny97

Differential Revision: D83356565

fbshipit-source-id: 16815bd76a33cc7f867c4361568c4280b5668390