Explain the existing Phi optimization better
Summary:
The existing code is slightly confusing - the optimization
to eliminate Phi bindings interacts with the optimization in
`add_branch` in a way that I suspect no one really had internalized.

This isn't really a problem today, but a refactor that tracks both narrows and
values isn't going to be able to optimize `add_branch`, so we need to rework
this, and it's helpful to first have a comment clearly explaining what happens.

Reviewed By: yangdanny97

Differential Revision: D83353405

fbshipit-source-id: 5b8ab16a8b555febac53145b196d5be205960470