Rename Key / Binding variants used in partial-type inference
Summary:
I never liked focusing on "pin" much for these names, because
"pin" is a verb that represents an action we take in other
circumstances (e.g. recursive Var) that aren't relevant to
what's being discussed here, which is partial type inference.

Changing the names to use a clear-cut noun phrase better aligns
them with how I want to be able to talk about this concept when
discussing the design and its maintenence going forward.

The name `PartialTypeWithUpstreamsCompleted` is a mouthful, but
I think that's okay - the action of the binding can't really
be described concisely, and anyone actually working with this
binding really needs to understand what it does.

Reviewed By: yangdanny97

Differential Revision: D83362522

fbshipit-source-id: 41036e2253ad2df7b07fc5784f6fcc0f05baa09a