Move `FlowInfo` up
Summary:
I find it easier to read code where the data structure defs either flow from small
to big (like OCaml) or from big to small (as is more often conventional in other
languages).

This code was inconsistently mixing both, and I need to layer in some additional
complexity. I generally prefer top-down unless I'm in a language that doesn't
allow it, so reorganizing that way (so that we go Flow -> FlowInfo -> FlowStyle).

Reviewed By: yangdanny97

Differential Revision: D83480569

fbshipit-source-id: 26489981cc7b339b7b87503851d1c381e42c3fe0