Increase gas
Summary:
To avoid runaway non-termination, we limit the work that subset can do. The default of 25 is too low, though. We might reasonably expect to hit that for complicated constraints.

We should not expect to hit this limit in practice, so a higher limit should always be okay. Increasing to 1000 for now.

Reviewed By: stroxler

Differential Revision: D83363337

fbshipit-source-id: 271778591257d4e83e8d1ade2dfc015038b59ab2