Unify vars with union-find
Summary:
Before this diff, we unified (Var, Var) by setting one var's answer to the other var. This diff uses union-find instead.

The variables map now contains Goto and Root nodes. When we look up a var, we walk over Goto nodes until we find a root. Unification of two roots turns one into a Goto pointing to the other.

Unification of vars is destructive -- unifying a lower-bound Quantified var with an upper-bound Contained var would "solve" the Quantified to the Contained, erasing the Quantified information. For now, this diff preserves this behavior (see comment on `unify` function).

Implementation-wise, I am making liberal use of interior mutability (RefCell and Cell). I have documented _how_ this interior mutability should be used, but note that the RefCell invariants are enforced at runtime, so the solver code is a bit trickier to write correctly after this change.

Reviewed By: stroxler

Differential Revision: D83374460

fbshipit-source-id: fa8e950c7d715e3bdfa30b8bb963fec45735dc46