Introduce unification-aware VarRecurser
Summary:
Later in this stack, I implement union-find for variables, which means that a given Var might just be a goto node. When recursing, we always want to track root nodes, so we need to find the root node first.

This diff adds a VarRecurser type which only exposes it's "recurse" function via Solver. We don't yet have the union-find data structure yet. This diff just adds the abstraction, and later we will add code to find the root var.

Reviewed By: stroxler

Differential Revision: D83374462

fbshipit-source-id: 7b3531ffd76299778be2f54ec714f05f6365a542