Separate variable insert and update
Summary:
I want to be sure that we never overwrite an existing var accidentally via insert. For example, if we overwrite a Goto node we could inadvertently change the identify of other vars.

This diff separates the existing `insert` method into separate `insert_fresh` and `update` functions, each of which panics if used incorrectly.

Reviewed By: stroxler

Differential Revision: D83374459

fbshipit-source-id: 1c7de29c14113c6e34b3b8de373a2b92451872b7