Use the pyrefly ast visitor instead of our own
Summary: Now that we moved the logic for global variables out of the `ast_visitor.rs` module, now that module is only responsible for exporting type of expressions. It only needs to iterate on all expressions, regardless of scope. Therefore, we can use the pyrefly AST visitor to make that implementation a lot simpler.

Reviewed By: stroxler

Differential Revision: D83479451

fbshipit-source-id: c412ec0e87b918b77b47c7c1b6001656a205cc1e