Perform inconsistent override check on more dunder methods
Summary:
When going through the list of "issues in Pyre1 that are fixed in Pyrefly", I noticed [one case](https://fb.workplace.com/groups/pyreqa/permalink/9482986728457741/) where Pyrefly currently preserves the Pyre1 bug of skipping inconsistent override checking for all dunder methods.

Some code archeology suggests that the behavior was first introduced [here](https://www.internalfb.com/diff/D69570486?dst_version_fbid=591340110559792&transaction_fbid=1319302492653411). So this is less of an intentional exclusion and more of a TODO that hasn't been addressed yet.

This diff attempts to address the issue, by skipping the inconsistent override check less aggressively. Previously, we skip all attributes whose names are in the form `_XXX_` and `__YYY`. Now, we start checking them and only skip the methods when the name is included in a known blocklist. This blocklist includes the following names:
- `__new__`, `__init__`. These two are theoretically not part of the interface of the classes and therefore should never be checked for overrides.
- `__call__`. This should be checked in theory, but there are practical implications (see code comments) that prevents us from turning it on right now.
- `__post_init__`. This is dataclass specific and doesn't follow normal class semantics.
- `_ignore_`. This is enum specific and doesn't follow normal class semantics.

Reviewed By: rchen152

Differential Revision: D83597969

fbshipit-source-id: f13344b37c6e3a90528f5c96355db55ad42968a9