Optimize `merge_flow`
Summary:
There's an unnecessary check here: we no longer need to merge a single flow
when it isn't terminated, because we are no longer only passing nonterminated
flows as we were when D65495213 was written.

Note that I've expanded the regression test so that it's Phi node isn't
optimized away, and I've added a comment about what the test is there to guard
against.

Reviewed By: grievejia

Differential Revision: D83516083

fbshipit-source-id: 9b609c04e7bd046bb717258cf80990e567d608a4