Change InitializedInFlow to UninitializedInFlow
Summary:
The existing name emphasizes the wrong thing - we don't really know
whether, for example, narrows are initialized.

We're looking for things that we *know* are uninitialized, hence the name
ought to reflect that more clearly. And the `Maybe` variant doesn't really
mean "we don't know", it means "we know there are some conditional flows
where the name is unbound", so `Conditionally` is a better name.

Reviewed By: grievejia

Differential Revision: D83516081

fbshipit-source-id: 33f07da670b9ab334457f11dae538817c7f3eb8a