Add a test case of bad merging logic for narrows
Summary:
If a name is a capture, we can have it not exist in the base flow, get
narrows in branches, and then we drop its original type and unsoundly
narrow it downstream.

This doesn't happen if we narrow in the branch tests because we negate
the tests in the base flow (hence `reveal_type(x)` gives an unsimplified
but correct type), but that is brittle - if the narrows happen anywhere
else we get the wrong answer (hence `reveal_type(y)` giving `B | C` when
we know `y` is definitely neither of those).

Reviewed By: grievejia

Differential Revision: D83506544

fbshipit-source-id: 19ef539bdba84b2aabb4b4e43f6ea2226fcaa85b