Provide smart accessors to flow info
Summary:
I've never loved that code three layers up in `Scopes` is accessing `flow.info`
as a raw map, but it hasn't been an issue; once we split narrows vs values though,
some of these calls have to be changed and it's much cleaner to do so if we are
using nicely layered APIs so that we can express what we really intend for value-only
operations.

This diff preps for that.

Reviewed By: grievejia

Differential Revision: D83516082

fbshipit-source-id: 0375d0655a8ebd7b267c9d3615f29eead0af2698