Add a method to ClassMetadata to get the metaclass even if it's `type`
Summary: Previously, the `metaclass` method returned the metaclass only if it wasn't `type`, which meant we had to separately handle `type` in several places. This diff renames `metaclass` to `custom_metaclass` to reflect that it only gets non-`type` metaclasses and adds a new `metaclass` method that can replace a lot of our duplicate `type` handling.

Reviewed By: grievejia

Differential Revision: D83625989

fbshipit-source-id: 090ddcbaee287872c79a52f73d1a22ebfe8e1ae3