Add `ProtocolSubset` variant to `AttributeBase1`
Summary:
Adds a new variant to `AttributeBase1` to track when an attribute lookup is being done as part of a protocol subset check. This diff adds the variant without yet doing anything with it. The next diff will use it to tweak how magic dunder method lookups are done for protocol checks, fixing an enum iteration bug.

For https://github.com/facebook/pyrefly/issues/1009.

Reviewed By: migeed-z

Differential Revision: D83624891

fbshipit-source-id: 80e1ae7a94aff776311ce67836de9d09fcd19e83