Implement a scope-aware AST visitor
Summary:
This diff implements a scope-aware AST visitor.
The visitor state contains a stack of scopes, which can be used to tell the current function, including wether it is pysa-exported or not.

This is then used to re-implement the call graph visitor.
In a future diff, it will also be used to compute captured variables.

Reviewed By: stroxler

Differential Revision: D83566870

fbshipit-source-id: cf7ccf16a15a54fa07639ed294bca038865cfcf0