Simplify `MergeItem`
Summary:
Merging needs to become more complex to optimize away all the Phi bindings
we want to. The current logic is more complex than it needs to be because
`MergeItem` is encapsulating things that can pretty easily be tracked
elsewhere, and I'd prefer to simplify and remove a bit of the indirection in
the current setup before adding more complexity.

Reviewed By: grievejia

Differential Revision: D83675224

fbshipit-source-id: 09012cefbaa3b264de52ac7fd5bf014f14009f8f