Eliminate `MergeItem`
Summary:
This diff eliminates the `MergeItem` datatype, which wasn't giving us much
anymore (it was really just a wrapper around `Vec<FlowInfo>`, and we have a
better, smarter wrapper that can manage insertions at a higher level in
`MergeItems`).

The resulting code is more direct with no callbacks, and puts the really
essential logic in `merged_flow_info` and `merge_flow` next to one another.

There are some more cleanups possible which I'll deal with in the next diff.

Reviewed By: grievejia

Differential Revision: D83675223

fbshipit-source-id: ca644833d4edd576090431e9e47787263d3010fd