Optimize out narrow-only branches in loop Phi
Summary:
In a loop, a branch that does nothing but narrow the binding set at the
top of the loop (the loop phi, which we store as a "value" binding)
can be ignored in flow merging.

The reason for this is that the merged flow includes the original binding
from above the loop, for which the loop phi is just a proxy in case there
are assignments.

We know that merging a narrow with the original type should generally result in
the original type.
- Pyrefly doesn't have narrows, and our flow merging is too naive to account for
  this, so in our current implementation this isn't true, but that's
  a limitation of Pyrefly, not anything inherent
- In semantic type theory this is absolutely true: a narrow is an intesection,
  and unioning the intersection of a set with the original gives back the
  original

If we put these two things together, it's clear that in a loop, any branch that
just narrows the phi isn't doing anything useful and can be dropped, regardless
of whether there are active narrows.

Note that there are two classes of loop-recursive-var bugs that are *not* fixed:
- Because we don't optimize the phi ahead of time, we still lose track of
  first-usage pinning for partial types in loops, which has been reported at
  least twice. I'm not sure whether we can fix this without doing an extra
  traversal of the AST before binding, and I'm not prepared to get into it
  at this time.
- In the case where there is assignment in the loop, we behave as before; in
  some cases it works, but in any complicated case we will probably produce
  bad errors, and even many trivial cases behave very badly because pinning
  eagerly is the wrong approach. Solving that case requires more than just
  optimizing phi, we must be able to actually model the phi in some new way.
- At the moment, we aren't modeling narrows vs values in nested control flow
  Phi; as a result, narrowing in nested control structures would break the
  optimization. We'll need to make the merging logic more complex before
  we can fix that.

Fixes https://github.com/facebook/pyrefly/issues/683

Reviewed By: grievejia

Differential Revision: D83675222

fbshipit-source-id: 880394ab5041db303f679e26eea39258704c2602